Final Movement complete
I had to redo the whole camera and movement system due to the player being able to walk through walls ( I did NOT make  all of the code I watch some videos and altered them ). I have also created a semi entrance and added fog for some that one word I cant spell. I THINK that is all I have done I also tested using a Terrain generator and fed it a image of a map but it was not ideal so I think it would be best to create it ourselves.