allot of small - medium size changes
Added some decoration outside of the maze ( can be changed / removed if its to much or does not fit the game ). added a timer. added some Ui for the timer and score ( I have not gotten the score to work so i need to figure out a way to do that). I have added a main menu scene with workign button to start the game, close the game and a how to play with some basic text for now. there is alot of placed objects now along side particls that last 10s but this might cause issues for your pc if its to much so  good luck