traps/scenes
I have added in most traps with the current layout of the maze, I will adjust these and add more when the maze is done. I have added a lot more walls that will dissappear and reappear at times to confuse the player. I have also created the win and Loose screen that all work properly. I have also models a very simple arrow that is used instead of a sphere and they dissapper shortly after they are spawned in.